backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.0.0
langgraph==0.0.20
openai==1.3.0
chromadb==0.4.18
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.24.0,<0.25.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1

